refactor(gateway): move `ValidateGatewayInterceptor` to device module
